refactor: ShoppingPage 페이지 로직 분리
