+++ src/basic/widgets/ShoppingSidebar/ui/index.ts
@@ -1,0 +1,1 @@
+export { ShoppingSidebar } from './ShoppingSidebar';
